+++ core/io/zip_io.cpp
@@ -102,8 +102,8 @@ int zipio_testerror(voidpf opaque, voidpf stream) {
 }
 
 voidpf zipio_alloc(voidpf opaque, uInt items, uInt size) {
-	voidpf ptr = memalloc(items * size);
-	zeromem(ptr, items * size);
+	voidpf ptr = memalloc((size_t)items * size);
+	zeromem(ptr, (size_t)items * size);
 	return ptr;
 }
