+++ scene/3d/gpu_particles_collision_3d.cpp
@@ -478,7 +478,7 @@ Ref<Image> GPUParticlesCollisionSDF::bake() {
 	_create_bvh(bvh, face_pos.ptr(), face_pos.size(), faces.ptr(), th);
 
 	Vector<uint8_t> data;
-	data.resize(sdf_size.z * sdf_size.y * sdf_size.x * sizeof(float));
+	data.resize((uint64_t)sdf_size.z * sdf_size.y * sdf_size.x * sizeof(float));
 
 	if (bake_step_function) {
 		bake_step_function(0, "Baking SDF");
